+++ src/types/interfaces/SpinnerOptions.ts
@@ -1,0 +1,7 @@
+import { Dictionary } from './Dictionary';
+
+export interface SpinnerOptions {
+  enabled?: boolean;
+  fillColor?: string;
+  styles?: Dictionary;
+}