+++ src/types/interfaces/index.ts
@@ -4,3 +4,4 @@ export { Options } from './Options';
 export { RawOptions } from './RawOptions';
 export { RawOptionsCallbacks } from './RawOptionsCallbacks';
 export { Position } from './Position';
+export { SpinnerOptions } from './SpinnerOptions';