+++ src/FilterContainer/styles.ts
@@ -1,16 +1,16 @@
 import FilterizrOptions from '../FilterizrOptions';
 
-export const makeInitialStyles = (options: FilterizrOptions): object => ({
+export const makePaddingStyles = (options: FilterizrOptions): object => ({
   padding: `${options.get().gutterPixels}px`,
+});
+
+export const makeInitialStyles = (options: FilterizrOptions): object => ({
+  ...makePaddingStyles(options),
   position: 'relative',
   // Needed for flex displays
   width: '100%',
   display: 'flex',
   flexWrap: 'wrap',
-});
-
-export const makePaddingStyles = (options: FilterizrOptions): object => ({
-  padding: `0 ${options.get().gutterPixels / 2}px`,
 });
 
 export const makeHeightStyles = (height: number): object => ({